feat: Create app, update data, and format code

+++ forest_fire.py
@@ -4,6 +4,7 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.model_selection import train_test_split
 import warnings
 import pickle
+
 warnings.filterwarnings("ignore")
 
 data = pd.read_csv("Forest_fire.csv")
@@ -11,8 +12,8 @@ data = np.array(data)
 
 X = data[1:, 1:-1]
 y = data[1:, -1]
-y = y.astype('int')
-X = X.astype('int')
+y = y.astype("int")
+X = X.astype("int")
 # print(X,y)
 X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=0)
 log_reg = LogisticRegression()
@@ -20,11 +21,11 @@ log_reg = LogisticRegression()
 
 log_reg.fit(X_train, y_train)
 
-inputt=[int(x) for x in "45 32 60".split(' ')]
-final=[np.array(inputt)]
+inputt = [int(x) for x in "45 32 60".split(" ")]
+final = [np.array(inputt)]
 
 b = log_reg.predict_proba(final)
 print(b)
 
-pickle.dump(log_reg,open('model.pkl','wb'))
+pickle.dump(log_reg, open("model.pkl", "wb"))
 # model=pickle.load(open('model.pkl','rb'))
